backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-jose[cryptography]
httpx
requests
pgvector
sentence-transformers
apscheduler
sendgrid
python-dotenv
alembic
numpy
pytest
pytest-asyncio
pytest-cov
gunicorn